AudienceSutra/requirements.txt
certifi==2018.8.24
chardet==3.0.4
defusedxml==0.5.0
Django==1.11
django-allauth==0.37.1
django-bootstrap-form==3.4
django-extensions==2.1.2
django-markup==1.2
django-otp==0.5.0
django-rest-auth==0.9.3
django-tables2==1.21.2
django-widget-tweaks==1.4.2
djangorestframework==3.8.2
idna==2.7
oauthlib==2.1.0
payu-biz==2.0
Pillow==5.1.0
pkg-resources==0.0.0
psycopg2==2.7.5
psycopg2-binary==2.7.5
python3-openid==3.1.0
pytz==2018.5
razorpay==1.2.0
requests==2.19.1
requests-oauthlib==1.0.0
six==1.11.0
urllib3==1.23
